notification-service/requirements.txt
fastapi
uvicorn[standard]
kafka-python
prometheus-client
pymongo
